+++ packages/evokit-list/src/index.js
@@ -2,11 +2,12 @@ import { createBlock } from 'evokit';
 
 export const List = createBlock('ul', 'list', [
     'color',
+    'display',
     'divider',
-    'style',
     'indent',
+    'style',
 ]);
 
-export const ListItem = createBlock('li', 'list__item');
+export const ListItem = createBlock('li', 'list__item', ['display']);
 
 List.Item = ListItem;